fix error codes to be working fully - on order and po line update
