+++ src/common/constants/errorCodes.js
@@ -1,0 +1,59 @@
+export const ERROR_CODES = {
+  accessProviderIsInactive: 'accessProviderIsInactive',
+  accessProviderNotFound: 'accessProviderNotFound',
+  budgetIsInactive: 'budgetIsInactive',
+  budgetNotFoundForTransaction: 'budgetNotFoundForTransaction',
+  costAdditionalCostInvalid: 'costAdditionalCostInvalid',
+  costDiscountInvalid: 'costDiscountInvalid',
+  costQtyPhysicalExceedsLoc: 'costQtyPhysicalExceedsLoc',
+  costUnitPriceElectronicInvalid: 'costUnitPriceElectronicInvalid',
+  costUnitPriceInvalid: 'costUnitPriceInvalid',
+  currentFYearNotFound: 'currentFYearNotFound',
+  electronicLocCostQtyMismatch: 'electronicLocCostQtyMismatch',
+  encumbranceCreationFailure: 'encumbranceCreationFailure',
+  fundCannotBePaid: 'fundCannotBePaid',
+  fundsNotFound: 'fundsNotFound',
+  idMismatch: 'idMismatch',
+  InstanceIdNotAllowedForPackagePoLine: 'InstanceIdNotAllowedForPackagePoLine',
+  invalidISBN: 'invalidISBN',
+  itemCreationFailed: 'itemCreationFailed',
+  itemNotFound: 'itemNotFound',
+  itemNotRetrieved: 'itemNotRetrieved',
+  itemUpdateFailed: 'itemUpdateFailed',
+  ledgerNotFoundForTransaction: 'ledgerNotFoundForTransaction',
+  locNotProvided: 'locNotProvided',
+  locQtyElectronicExceedsCost: 'locQtyElectronicExceedsCost',
+  locQtyPhysicalExceedsCost: 'locQtyPhysicalExceedsCost',
+  materialTypeRequired: 'materialTypeRequired',
+  missingContributorNameType: 'missingContributorNameType',
+  missingInstanceStatus: 'missingInstanceStatus',
+  missingInstanceType: 'missingInstanceType',
+  missingLoanType: 'missingLoanType',
+  missingOngoing: 'missingOngoing',
+  nonZeroCostQtyElectronic: 'nonZeroCostQtyElectronic',
+  nonZeroCostQtyPhysical: 'nonZeroCostQtyPhysical',
+  nonZeroLocQtyPhysical: 'nonZeroLocQtyPhysical',
+  ongoingNotAllowed: 'ongoingNotAllowed',
+  orderAcqUnitsNotFound: 'orderAcqUnitsNotFound',
+  orderApprovalRequired: 'orderApprovalRequired',
+  orderClosed: 'orderClosed',
+  orderIdMismatch: 'orderIdMismatch',
+  orderIdRequired: 'orderIdRequired',
+  orderNotFound: 'orderNotFound',
+  orderOpen: 'orderOpen',
+  organizationNotAVendor: 'organizationNotAVendor',
+  physicalLocCostQtyMismatch: 'physicalLocCostQtyMismatch',
+  poNumberNotUnique: 'poNumberNotUnique',
+  poNumberRequired: 'poNumberRequired',
+  protectedFieldChanging: 'protectedFieldChanging',
+  userHasNoAcqUnitsPermission: 'userHasNoAcqUnitsPermission',
+  userHasNoApprovalPermission: 'userHasNoApprovalPermission',
+  userHasNoPermission: 'userHasNoPermission',
+  vendorIsInactive: 'vendorIsInactive',
+  vendorIssue: 'vendorIssue',
+  vendorNotFound: 'vendorNotFound',
+  zeroCostQty: 'zeroCostQty',
+  zeroCostQtyElectronic: 'zeroCostQtyElectronic',
+  zeroCostQtyPhysical: 'zeroCostQtyPhysical',
+  zeroLocQty: 'zeroLocQty',
+};
