+++ translations/ui-orders/en.json
@@ -349,6 +349,7 @@
   "errors.accessProviderNotFound": "POL {poLineNumber} is using an Access Provider that is inactive or has been removed, please select an active Access Provider to continue.",
   "errors.budgetIsInactive": "Cannot create encumbrance from the not active budget",
   "errors.budgetNotFoundForTransaction": "One or more fund distribution(s) on this order can not be encumbered, because there is no current budget",
+  "errors.clone.error": "The purchase order was not cloned",
   "errors.costAdditionalCostInvalid": "Cost's additional cost must be positive number",
   "errors.costDiscountInvalid": "Cost's discount is invalid",
   "errors.costQtyPhysicalExceedsLoc": "Cost's physical quantity exceeds locations' physical quantity",
@@ -519,7 +520,6 @@
   "order.lastUpdated": "Last updated",
   "order.acquisitionsUnit": "Acquisitions unit",
   "order.clone.confirmLabel": "Clone",
-  "order.clone.error": "The purchase order was not cloned",
   "order.clone.heading": "Clone order?",
   "order.clone.message": "Are you sure you want to clone this purchase order and all of its purchase order lines?",
   "order.clone.success": "The purchase order was successfully cloned",
